Aggregate source data

Aggregates sourcedata object based on fieldtypes specified at load time. Default time period for aggregation is a calendar day

Usage

aggregate_data(sourcedata, aggregation_timeunit = "day", showprogress = TRUE)

Arguments

sourcedata

A sourcedata object returned from prepare_data function

aggregation_timeunit

Unit of time to aggregate over. Specify one of "day", "week", "month", "quarter", "year". The "week" option is Monday-based. Default = "day"

showprogress

Print progress to console. Default = TRUE

Value

An aggregatedata object

See also

Examples

# load example data into a data.frame
rawdata <- read_data(
  system.file("extdata", "example_data.csv", package = "daiquiri"),
  delim = ",",
  col_names = TRUE
)

# validate and prepare the data for aggregation
sourcedataobj <- prepare_data(
  rawdata,
  fieldtypes = fieldtypes(PrescriptionID = ft_uniqueidentifier(),
    PrescriptionDate = ft_timepoint(),
    AdmissionDate = ft_datetime(includes_time = FALSE),
    Drug = ft_freetext(),
    Dose = ft_numeric(),
    DoseUnit = ft_categorical(),
    PatientID = ft_ignore(),
    Location = ft_categorical(aggregate_by_each_category=TRUE)),
  override_columnnames = FALSE,
  na = c("","NULL")
)
#> Fieldtypes supplied:
#> PrescriptionID	<fieldtype_uniqueidentifier>
#> PrescriptionDate	<fieldtype_timepoint>	options: includes_time
#> AdmissionDate	<fieldtype_datetime>
#> Drug	<fieldtype_freetext>
#> Dose	<fieldtype_numeric>
#> DoseUnit	<fieldtype_categorical>
#> PatientID	<fieldtype_ignore>
#> Location	<fieldtype_categorical>	options: aggregate_by_each_category
#>  
#> Checking column names against fieldtypes... 
#> Importing source data [NULL]... 
#> Checking data against fieldtypes... 
#>   Selecting relevant warnings... 
#>   Identifying nonconformant values... 
#>   Checking and removing missing timepoints... 
#> Checking for duplicates... 
#>   Sorting data... 
#> Loading into sourcedata structure... 
#>   PrescriptionID 
#>   PrescriptionDate 
#>   AdmissionDate 
#>   Drug 
#>   Dose 
#>   DoseUnit 
#>   PatientID 
#>   Location 
#> Finished 

# aggregate the data
aggregatedataobj <- aggregate_data(
  sourcedataobj,
  aggregation_timeunit = "day"
)
#> Aggregating [] by [day]... 
#> Aggregating overall dataset... 
#> Aggregating each datafield in turn... 
#> 1: PrescriptionID 
#> Preparing... 
#> Aggregating character field... 
#>   By n 
#>   By missing_n 
#>   By missing_perc 
#>   By minlength 
#>   By maxlength 
#>   By meanlength 
#> Finished 
#> 2: PrescriptionDate 
#> Preparing... 
#> Aggregating double field... 
#>   By n 
#>   By midnight_n 
#>   By midnight_perc 
#> Finished 
#> 3: AdmissionDate 
#> Preparing... 
#> Aggregating double field... 
#>   By n 
#>   By missing_n 
#>   By missing_perc 
#>   By nonconformant_n 
#>   By nonconformant_perc 
#>   By min 
#>   By max 
#> Finished 
#> 4: Drug 
#> Preparing... 
#> Aggregating character field... 
#>   By n 
#>   By missing_n 
#>   By missing_perc 
#> Finished 
#> 5: Dose 
#> Preparing... 
#> Aggregating double field... 
#>   By n 
#>   By missing_n 
#>   By missing_perc 
#>   By nonconformant_n 
#>   By nonconformant_perc 
#>   By min 
#>   By max 
#>   By mean 
#>   By median 
#> Finished 
#> 6: DoseUnit 
#> Preparing... 
#> Aggregating character field... 
#>   By n 
#>   By missing_n 
#>   By missing_perc 
#>   By distinct 
#> Finished 
#> 7: Location 
#> Preparing... 
#> Aggregating character field... 
#>   By n 
#>   By missing_n 
#>   By missing_perc 
#>   By distinct 
#>   By subcat_n 
#>     2 categories found 
#>     1: SITE1 
#>     2: SITE2 
#>   By subcat_perc 
#>     2 categories found 
#>     1: SITE1 
#>     2: SITE2 
#> Finished 
#> Aggregating calculated fields... 
#> [DUPLICATES]: 
#> Preparing... 
#> Aggregating integer field... 
#>   By sum 
#>   By nonzero_perc 
#> Finished 
#> [ALLFIELDSCOMBINED]: 
#> Finished