Generate report from existing objects

Generate report from previously-created daiquiri_source_data and daiquiri_aggregated_data objects

Usage

report_data(
  source_data,
  aggregated_data,
  report_title = "daiquiri data quality report",
  save_directory = ".",
  save_filename = NULL,
  format = "html",
  show_progress = TRUE,
  ...
)

Arguments

source_data

A daiquiri_source_data object returned from prepare_data() function

aggregated_data

A daiquiri_aggregated_data object returned from aggregate_data() function

report_title

Title to appear on the report

save_directory

String specifying directory in which to save the report. Default is current directory.

save_filename

String specifying filename for the report, excluding any file extension. If no filename is supplied, one will be automatically generated with the format daiquiri_report_YYMMDD_HHMMSS.

format

File format of the report. Currently only "html" is supported

show_progress

Print progress to console. Default = TRUE

...

Further parameters to be passed to rmarkdown::render(). Cannot include any of input, output_dir, output_file, params, quiet.

Value

A string containing the name and path of the saved report

See also

Examples

# \donttest{
# load example data into a data.frame
raw_data <- read_data(
  system.file("extdata", "example_prescriptions.csv", package = "daiquiri"),
  delim = ",",
  col_names = TRUE
)

# validate and prepare the data for aggregation
source_data <- prepare_data(
  raw_data,
  field_types = field_types(
    PrescriptionID = ft_uniqueidentifier(),
    PrescriptionDate = ft_timepoint(),
    AdmissionDate = ft_datetime(includes_time = FALSE),
    Drug = ft_freetext(),
    Dose = ft_numeric(),
    DoseUnit = ft_categorical(),
    PatientID = ft_ignore(),
    Location = ft_categorical(aggregate_by_each_category = TRUE)
  ),
  override_column_names = FALSE,
  na = c("", "NULL"),
  dataset_description = "Example data provided with package",
  show_progress = TRUE
)
#> field_types supplied:
#> PrescriptionID	<uniqueidentifier>
#> PrescriptionDate	<timepoint>	options: includes_time
#> AdmissionDate	<datetime>
#> Drug	<freetext>
#> Dose	<numeric>
#> DoseUnit	<categorical>
#> PatientID	<ignore>
#> Location	<categorical>	options: aggregate_by_each_category
#>  
#> Checking column names against field_types... 
#> Importing source data [Example data provided with package]... 
#> Checking data against field_types... 
#>   Selecting relevant warnings... 
#>   Identifying nonconformant values... 
#>   Checking and removing missing timepoints... 
#> Checking for duplicates... 
#>   Sorting data... 
#> Loading into source_data structure... 
#>   PrescriptionID 
#>   PrescriptionDate 
#>   AdmissionDate 
#>   Drug 
#>   Dose 
#>   DoseUnit 
#>   PatientID 
#>   Location 
#> Finished 

# aggregate the data
aggregated_data <- aggregate_data(
  source_data,
  aggregation_timeunit = "day",
  show_progress = TRUE
)
#> Aggregating [] by [day]... 
#> Aggregating overall dataset... 
#> Aggregating each data_field in turn... 
#> 1: PrescriptionID 
#> Preparing... 
#> Aggregating character field... 
#>   By n 
#>   By missing_n 
#>   By missing_perc 
#>   By min_length 
#>   By max_length 
#>   By mean_length 
#> Finished 
#> 2: PrescriptionDate 
#> Preparing... 
#> Aggregating double field... 
#>   By n 
#>   By midnight_n 
#>   By midnight_perc 
#> Finished 
#> 3: AdmissionDate 
#> Preparing... 
#> Aggregating double field... 
#>   By n 
#>   By missing_n 
#>   By missing_perc 
#>   By nonconformant_n 
#>   By nonconformant_perc 
#>   By min 
#>   By max 
#> Finished 
#> 4: Drug 
#> Preparing... 
#> Aggregating character field... 
#>   By n 
#>   By missing_n 
#>   By missing_perc 
#> Finished 
#> 5: Dose 
#> Preparing... 
#> Aggregating double field... 
#>   By n 
#>   By missing_n 
#>   By missing_perc 
#>   By nonconformant_n 
#>   By nonconformant_perc 
#>   By min 
#>   By max 
#>   By mean 
#>   By median 
#> Finished 
#> 6: DoseUnit 
#> Preparing... 
#> Aggregating character field... 
#>   By n 
#>   By missing_n 
#>   By missing_perc 
#>   By distinct 
#> Finished 
#> 7: Location 
#> Preparing... 
#> Aggregating character field... 
#>   By n 
#>   By missing_n 
#>   By missing_perc 
#>   By distinct 
#>   By subcat_n 
#>     2 categories found 
#>     1: SITE1 
#>     2: SITE2 
#>   By subcat_perc 
#>     2 categories found 
#>     1: SITE1 
#>     2: SITE2 
#> Finished 
#> Aggregating calculated fields... 
#> [DUPLICATES]: 
#> Preparing... 
#> Aggregating integer field... 
#>   By sum 
#>   By nonzero_perc 
#> Finished 
#> [ALL_FIELDS_COMBINED]: 
#> Finished 

# save a report in the current directory using the previously-created objects
report_data(
  source_data,
  aggregated_data,
  report_title = "daiquiri data quality report",
  save_directory = ".",
  save_filename = "example_data_report",
  show_progress = TRUE
)
#> Generating html report... 
#> 
#> 
#> processing file: report_htmldoc.Rmd
#>   |                                                                              |                                                                      |   0%  |                                                                              |..                                                                    |   3%
#>    inline R code fragments
#> 
#>   |                                                                              |.....                                                                 |   7%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                              |.......                                                               |  10%
#>   ordinary text without R code
#> 
#>   |                                                                              |..........                                                            |  14%
#> label: unnamed-chunk-1 (with options) 
#> List of 2
#>  $ echo  : logi FALSE
#>  $ engine: chr "css"
#> 
#>   |                                                                              |............                                                          |  17%
#>    inline R code fragments
#> 
#>   |                                                                              |..............                                                        |  21%
#> label: source_data
#>   |                                                                              |.................                                                     |  24%
#>   ordinary text without R code
#> 
#>   |                                                                              |...................                                                   |  28%
#> label: fields-imported
#>   |                                                                              |......................                                                |  31%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................                                              |  34%
#> label: fields-ignored
#>   |                                                                              |...........................                                           |  38%
#>   ordinary text without R code
#> 
#>   |                                                                              |.............................                                         |  41%
#> label: validation-warnings
#>   |                                                                              |...............................                                       |  45%
#>   ordinary text without R code
#> 
#>   |                                                                              |..................................                                    |  48%
#> label: source_data-summary
#>   |                                                                              |....................................                                  |  52%
#>   ordinary text without R code
#> 
#>   |                                                                              |.......................................                               |  55%
#> label: aggregated_data
#>   |                                                                              |.........................................                             |  59%
#>   ordinary text without R code
#> 
#>   |                                                                              |...........................................                           |  62%
#> label: overview-presence
#>   |                                                                              |..............................................                        |  66%
#>   ordinary text without R code
#> 
#>   |                                                                              |................................................                      |  69%
#> label: overview-missing
#>   |                                                                              |...................................................                   |  72%
#>   ordinary text without R code
#> 
#>   |                                                                              |.....................................................                 |  76%
#> label: overview-nonconformant
#>   |                                                                              |........................................................              |  79%
#>   ordinary text without R code
#> 
#>   |                                                                              |..........................................................            |  83%
#> label: overview-duplicates
#>   |                                                                              |............................................................          |  86%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............................................................       |  90%
#> label: aggregated_data-summary
#>   |                                                                              |.................................................................     |  93%
#>   ordinary text without R code
#> 
#>   |                                                                              |....................................................................  |  97%
#> label: individual-fields (with options) 
#> List of 1
#>  $ results: chr "asis"
#> 
#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 
#> output file: report_htmldoc.knit.md
#> "C:/Program Files/RStudio/bin/pandoc/pandoc" +RTS -K512m -RTS report_htmldoc.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output pandoc402848aa3cba.html --lua-filter "C:\Users\phuongq\Documents\Rworkspace\library\rmarkdown\rmarkdown\lua\pagebreak.lua" --lua-filter "C:\Users\phuongq\Documents\Rworkspace\library\rmarkdown\rmarkdown\lua\latex-div.lua" --self-contained --variable bs3=TRUE --standalone --section-divs --template "C:\Users\phuongq\Documents\Rworkspace\library\rmarkdown\rmd\h\default.html" --no-highlight --variable highlightjs=1 --variable theme=bootstrap --mathjax --variable "mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML" --include-in-header "C:\Users\phuongq\AppData\Local\Temp\RtmpEXPt0Z\rmarkdown-str40283a2d804.html" 
#> 
#> Output created: example_data_report.html
#> Report saved to: ./example_data_report.html 
#> [1] "./example_data_report.html"
file.remove("./example_data_report.html")
#> [1] TRUE
# }