Prepare source data

Validate a data frame against a fieldtypes specification, and prepare for aggregation.

Usage

prepare_data(
  df,
  fieldtypes,
  override_columnnames = FALSE,
  na = c("", "NA", "NULL"),
  dataset_shortdesc = NULL,
  showprogress = TRUE
)

Arguments

df

A data frame

fieldtypes

fieldtypes object specifying names and types of fields (columns) in the source data. See also availablefieldtypes.

override_columnnames

If FALSE, column names must exist in data frame and must match the names specified in fieldtypes exactly. If TRUE, column names in source will be replaced with names in fieldtypes specification. The specification must therefore contain the columns in the correct order. Default = FALSE

na

vector containing strings that should be interpreted as missing values, Default = c("","NA","NULL").

dataset_shortdesc

Short description of the dataset being checked. This will appear on the report. If blank, the name of the data frame object will be used

showprogress

Print progress to console. Default = TRUE

Value

A sourcedata object

See also

Examples

rawdata <- read_data(
  system.file("extdata", "example_data.csv", package = "daiquiri"),
  delim = ",",
  col_names = TRUE
)

sourcedataobj <- prepare_data(
  rawdata,
  fieldtypes = fieldtypes(PrescriptionID = ft_uniqueidentifier(),
    PrescriptionDate = ft_timepoint(),
    AdmissionDate = ft_datetime(includes_time = FALSE),
    Drug = ft_freetext(),
    Dose = ft_numeric(),
    DoseUnit = ft_categorical(),
    PatientID = ft_ignore(),
    Location = ft_categorical(aggregate_by_each_category=TRUE)),
  override_columnnames = FALSE,
  na = c("","NULL")
)
#> Fieldtypes supplied:
#> PrescriptionID	<fieldtype_uniqueidentifier>
#> PrescriptionDate	<fieldtype_timepoint>	options: includes_time
#> AdmissionDate	<fieldtype_datetime>
#> Drug	<fieldtype_freetext>
#> Dose	<fieldtype_numeric>
#> DoseUnit	<fieldtype_categorical>
#> PatientID	<fieldtype_ignore>
#> Location	<fieldtype_categorical>	options: aggregate_by_each_category
#>  
#> Checking column names against fieldtypes... 
#> Importing source data [NULL]... 
#> Checking data against fieldtypes... 
#>   Selecting relevant warnings... 
#>   Identifying nonconformant values... 
#>   Checking and removing missing timepoints... 
#> Checking for duplicates... 
#>   Sorting data... 
#> Loading into sourcedata structure... 
#>   PrescriptionID 
#>   PrescriptionDate 
#>   AdmissionDate 
#>   Drug 
#>   Dose 
#>   DoseUnit 
#>   PatientID 
#>   Location 
#> Finished