Read data from delimited file

Read rectangular data from a delimited file, with all columns read in as character type so that the package can later detect any non-conformant values. Operates as a restricted implementation of readr::read_delim.

Usage

read_data(
  file,
  delim = NULL,
  col_names = TRUE,
  quote = "\"",
  trim_ws = TRUE,
  comment = "",
  skip = 0,
  n_max = Inf,
  showprogress = TRUE
)

Arguments

file

A string containing path of file containing data to load, or a URL starting http://, file://, etc. Compressed files with extension ‘.gz’, ‘.bz2’, ".xz" and ".zip" are supported.

delim

Single character used to separate fields within a record. E.g. "," or "\t"

col_names

Either TRUE, FALSE or a character vector of column names. If TRUE, the first row of the input will be used as the column names, and will not be included in the data frame. If FALSE, column names will be generated automatically: X1, X2, X3 etc. Default = TRUE

quote

Single character used to quote strings.

trim_ws

Should leading and trailing whitespace be trimmed from each field?

comment

A string used to identify comments. Any text after the comment characters will be silently ignored

skip

Number of lines to skip before reading data. If comment is supplied any commented lines are ignored after skipping

n_max

Maximum number of lines to read.

showprogress

Display a progress bar? Default = TRUE

Value

A data frame

See also

Examples

rawdata <- read_data(
  system.file("extdata", "example_data.csv", package = "daiquiri"),
  delim = ",",
  col_names = TRUE
)